app/requirements.txt
gradio>=4.0.0
torch
transformers>=5.1.0
accelerate
pillow
opencv-python-headless
numpy